+++ src/main/java/roomescape/user/controller/ReservationPageController.java
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import roomescape.theme.ThemeService;
 import roomescape.user.service.ReservationService;
 import roomescape.user.service.ReservationTimeService;
-import roomescape.user.service.ThemeService;
 
 @Controller
 public class ReservationPageController {
