+++ src/test/java/roomescape/theme/controller/UserThemeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.user.controller;
+package roomescape.theme.controller;
 
 import static org.hamcrest.Matchers.is;
 
@@ -9,10 +9,29 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ThemeControllerTest {
+public class UserThemeControllerTest {
 
     @Test
-    void 사용자_트렌드_테마_조회() throws Exception {
+    void 전체_테마_조회() throws Exception {
+        RestAssured.given().log().all()
+                .when().get("/user/themes")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(11));
+
+    }
+
+    @Test
+    void 단일_테마_조회() {
+        RestAssured.given().log().all()
+                .when().get("/user/themes/1")
+                .then().log().all()
+                .statusCode(200)
+                .body("name", is("은하수"));
+    }
+
+    @Test
+    void 트렌드_테마_조회() throws Exception {
         RestAssured.given().log().all()
                 .when().get("/user/themes/trending?startDate=2026-05-01&endDate=2026-05-07&limit=10")
                 .then().log().all()
